refactor(remotestorage): move classes to separate files
Refs: SITKO-CORE-T-4
