+++ src/Sitko.Core.Storage.Remote.Server/BaseRemoteStorageController.cs
@@ -188,60 +188,3 @@ public abstract class BaseRemoteStorageController<TStorageOptions, TMetadata> : 
         return Ok(JsonSerializer.Serialize(result));
     }
 }
-
-public class UploadStorageItem<TMetadata>
-{
-    public IFormFile? File { get; set; }
-    public string? Path { get; set; }
-
-    [ModelBinder(BinderType = typeof(UploadStorageItemBinder))]
-    public TMetadata? Metadata { get; set; }
-
-    public string? FileName { get; set; }
-}
-
-public class UploadStorageItemBinder : IModelBinder
-{
-    private readonly ILogger<UploadStorageItemBinder> logger;
-
-    public UploadStorageItemBinder(ILogger<UploadStorageItemBinder> logger) => this.logger = logger;
-
-    public Task BindModelAsync(ModelBindingContext bindingContext)
-    {
-        if (bindingContext == null)
-        {
-            throw new ArgumentNullException(nameof(bindingContext));
-        }
-
-        var modelName = bindingContext.ModelName;
-
-        var valueProviderResult = bindingContext.ValueProvider.GetValue(modelName);
-
-        if (valueProviderResult == ValueProviderResult.None)
-        {
-            return Task.CompletedTask;
-        }
-
-        bindingContext.ModelState.SetModelValue(modelName, valueProviderResult);
-
-        var value = valueProviderResult.FirstValue;
-
-        if (string.IsNullOrEmpty(value))
-        {
-            return Task.CompletedTask;
-        }
-
-        try
-        {
-            var result = JsonSerializer.Deserialize(value, bindingContext.ModelType);
-            bindingContext.Result = ModelBindingResult.Success(result);
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, ex.Message);
-            bindingContext.Result = ModelBindingResult.Failed();
-        }
-
-        return Task.CompletedTask;
-    }
-}
