+++ src/Sitko.Core.Storage.Remote.Server/UploadStorageItemBinder.cs
@@ -1,0 +1,53 @@
+﻿using System;
+using System.Text.Json;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.Extensions.Logging;
+
+namespace Sitko.Core.Storage.Remote.Server;
+
+public class UploadStorageItemBinder : IModelBinder
+{
+    private readonly ILogger<UploadStorageItemBinder> logger;
+
+    public UploadStorageItemBinder(ILogger<UploadStorageItemBinder> logger) => this.logger = logger;
+
+    public Task BindModelAsync(ModelBindingContext bindingContext)
+    {
+        if (bindingContext == null)
+        {
+            throw new ArgumentNullException(nameof(bindingContext));
+        }
+
+        var modelName = bindingContext.ModelName;
+
+        var valueProviderResult = bindingContext.ValueProvider.GetValue(modelName);
+
+        if (valueProviderResult == ValueProviderResult.None)
+        {
+            return Task.CompletedTask;
+        }
+
+        bindingContext.ModelState.SetModelValue(modelName, valueProviderResult);
+
+        var value = valueProviderResult.FirstValue;
+
+        if (string.IsNullOrEmpty(value))
+        {
+            return Task.CompletedTask;
+        }
+
+        try
+        {
+            var result = JsonSerializer.Deserialize(value, bindingContext.ModelType);
+            bindingContext.Result = ModelBindingResult.Success(result);
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(ex, ex.Message);
+            bindingContext.Result = ModelBindingResult.Failed();
+        }
+
+        return Task.CompletedTask;
+    }
+}
